Basic character representation
Characters are CANNON bodies made of two boxes, one for the head and the
other for the body.
The body shape can change of height based on if the character is
crounched or standing up.
Mouse control is managed only client side, the server only correct
character movements to avoid cheating.